requirements.txt
Django == 4.2.3
psycopg2 == 2.9.6
djangorestframework == 3.14.0
djangorestframework_simplejwt == 5.2.2
django-model-utils == 4.3.1
django-cors-headers == 4.2.0